chore(): avoid puppeteer and playwright collisions
