+++ core/src/components/button/test/basic/button.e2e.p.ts
@@ -1,5 +1,5 @@
 import { expect, describe } from '@playwright/test';
-import { test } from '@utils/test/utils';
+import { test } from '@utils/test/utils-new';
 
 test.describe('button: basic', () => {
   test('should not have visual regressions', async ({ page }) => {
